+++ CHANGES.md
@@ -12,6 +12,9 @@
    ![Logo](logo-dark.png#dark-mode-only)
    ```
 
+* A pin icon is now used in the dataset selector to mark a dataset
+  that holds a pinned variable. (#424)
+  
 *  Slimmed down the main application bar by grouping features under a 
    single button that opens a dropdown menu. The features are: 
    Documentation, Developer Reference, Imprint, and Settings. (#540)